requirements.txt
av==10.0.0
accelerate==0.21.0
black==23.11.0
cloudpickle==2.2.1
clip==0.2.0
controlnet-aux==0.0.7
datasets==2.13.0
decord==0.6.0
diffusers==0.23.1
easyocr==1.7.1
easydict==1.11
einops==0.7.0
fairscale==0.4.13
fastapi==0.104.1
fire==0.5.0
ftfy==6.1.3
gradio==4.7.1
gradio_client==0.7.0
git+http://github.com/IDEA-Research/GroundingDINO.git
imageio==2.31.5
joblib==1.3.2
huggingface==0.0.1
huggingface-hub==0.17.3
langchain==0.0.348
Markdown==3.5.1
markdown-it-py==3.0.0
markdown2==2.4.11
matplotlib==3.8.0
mediapipe==0.10.8
modelscope==1.9.4
moviepy==1.0.3
munch==4.0.0
nltk==3.8.1
numpy==1.25.2
omegaconf==2.3.0
openai==1.3.7
openai-whisper==20230918
open-clip-torch==2.23.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
onnx==1.15.0
onnxruntime 
pandas==2.1.3
peft==0.4.0
psutil==5.9.5
pycocotools==2.0.7
pydantic==2.5.2
pydub==0.25.1
Pygments==2.16.1
PyYAML==6.0.1
pytorch_lightning==1.7.7
regex==2023.10.3
rotary-embedding-torch==0.4.0
scipy==1.11.4
soundfile==0.12.1
termcolor==2.4.0
tiktoken==0.3.3
timm==0.6.13
tqdm==4.66.1
transformers==4.34.1
torchmetrics==0.11.4
uvicorn==0.24.0.post1
xformers==0.0.22
wget==3.2
